==> R/00_all.qmd <==
---
title: "00_all"
format:
  html:
    embed-resources: true
editor: visual
---

# Load packages

```{r}
library("tidyverse")
library("quarto")
```

# Creating directories

```{r}
#dir.create("../data/_raw", recursive = TRUE)
#dir.create("../results")
#dir.create("../doc")
```

```{r}
#| warning: false
#| message: false
# Run 01_load file and save it as .html in the results folder
quarto::quarto_render("01_load.qmd")
file.rename("01_load.html", "../results/01_load.html")

# Run 02_clean file and save it as .html in the results folder
quarto::quarto_render("02_clean.qmd")
file.rename("02_clean.html", "../results/02_clean.html")

# Run 03_describe file and save it as .html in the results folder
quarto::quarto_render("03_describe.qmd")
file.rename("03_describe.html", "../results/03_describe.html")

# Run 04_augment file and save it as .html in the results folder
quarto::quarto_render("04_augment.qmd")
file.rename("04_augment.html", "../results/04_augment.html")

# Run 05_analysis_1 file and save it as .html in the results folder
quarto::quarto_render("05_analysis_1.qmd")
file.rename("05_analysis_1.html", "../results/05_analysis_1.html")

# Run 06_analysis_2 file and save it as .html in the results folder
quarto::quarto_render("06_analysis_2.qmd")
file.rename("06_analysis_2.html", "../results/06_analysis_2.html")

# Run 07_analysis_3 file and save it as .html in the results folder
quarto::quarto_render("07_analysis_3.qmd")
file.rename("07_analysis_3.html", "../results/07_analysis_3.html")


#Run Presentation file and save it as .html in the results folder
quarto::quarto_render("../doc/presentation.qmd")
```
