==> R/01_load.qmd <==
---
title: "01_load"
format:
  html:
    embed-resources: true
editor: visual
---

# Load packages

```{r}
library(tidyverse)
library(readr)
library(readxl)
```

# Setting the initial and final path

```{r}
#raw.githubusercontent.com is needed so we make sure that we are downloading the raw file
original_path <- "https://raw.githubusercontent.com/Masaam2000/Data_reexam/main/citotoxicity_genes_breast_cancer.csv"

destination_file <- "../data/_raw/citotoxicity_genes_breast_cancer.csv"
```

# Download the file

```{r}
download.file(original_path, 
              destination_file, 
              method = "auto")
```

# Transform the file from csv to tsv

```{r}
#This code separates the data with tabs and not comas
raw_data <- read_csv(destination_file)
raw_data <- write_delim(raw_data, 
                        file= "../data/_raw/citotoxicity_genes_breast_cancer.tsv",      
                        delim = "\t", 
                        col_names = TRUE)
```
