==> R/04_augment.qmd <==
---
title: "04_augment"
format:
  html:
    embed-resources: true
editor: visual
---

# Load packages

```{r Load the libraries}
library(tidyverse)
```

# Load the data

```{r Load the data}
setwd("C:/Users/marin/OneDrive/Documentos/GitHub/s223690/R")
res <- read.delim("../data/DESeq2_results.tsv",
                         row.names = 1, 
                         stringsAsFactors = FALSE)
```

# Check what are the deferentially expressed genes

The threshold used in this case is a strict Log2FoldChange of 3 and a p-adj value superior than 0.1. These values are very strict since we want to know the most deferentially expressed values. The normal values are a Log2FoldChange of 2 and a p-adj value superior than 0.05.

```{r Check the DE genes}
#Get the differentially expressed genes
de_genes <- subset(res, padj < 0.1)
de_genes <- subset(de_genes, log2FoldChange > 3 | log2FoldChange < -3)
de_genes <- subset(de_genes, !is.na(padj))
```

# Save data

```{r Save the DE genes}
# Define the path for the output .tsv file
output_file_path <- "../data/de_genes.tsv"

# Save the subset data frame to a .tsv file
write.table(de_genes, file = output_file_path, sep = "\t", quote = FALSE, row.names = TRUE)

# Verify that the file has been saved
if (file.exists(output_file_path)) {
  cat("DE genes has been saved to", output_file_path, "\n")
} else {
  cat("Failed to save DE genes in", output_file_path, "\n")
}

```
