==> R/06_analysis_2.qmd <==
---
title: "06_analysis_2"
format:
  html:
    embed-resources: true
editor: visual
---

# Load packages

```{r}
library(tidyverse)
library(quarto)
library(EnhancedVolcano)
```

# Load the data

```{r Load the data from 02_clean.qmd}
setwd("C:/Users/marin/OneDrive/Documentos/GitHub/s223690/R")
res <- read.delim("../data/DESeq2_results.tsv",
                         row.names = 1, 
                         stringsAsFactors = FALSE)
```

# Create a volcano plot

```{r Volcano plot of DE genes}
volcano_plot_data <- as.data.frame(res)

#See what are the maximum and minimum values for the plot
max_log2_volcano_plot <- max(volcano_plot_data$log2FoldChange)
min_log2_volcano_plot <- min(volcano_plot_data$log2FoldChange)

max_pvalue_volcano_plot <- max(-log10(volcano_plot_data$pvalue))
min_pvalue_volcano_plot <- min(-log10(volcano_plot_data$pvalue))


#Create the volcano plot (log2Fold change = 3, pvalue < 0.1)

#jpeg("../results/volcano_plot.png", width = 700, height = 500, res = 100)

EnhancedVolcano(volcano_plot_data,
                lab = row.names(volcano_plot_data),
                labSize = 4,
                x = "log2FoldChange",
                xlim = c(min_log2_volcano_plot, max_log2_volcano_plot),
                y = "padj",
                ylim = c(min_pvalue_volcano_plot, max_pvalue_volcano_plot),
                pCutoff = 0.1,
                FCcutoff = 3,
                axisLabSize = 12,
                caption = NULL,
                subtitle = NULL,
                legendPosition = "none",
                title = "Volcano plot"
                )
#dev.off()
```
