==> R/07_analysis_3.qmd <==
---
title: "07_analysis_3"
format:
  html:
    embed-resources: true
editor: visual
---

# Load packages

```{r}
library(clusterProfiler)
library(AnnotationDbi)
library(org.Hs.eg.db)
```

# Load the data

```{r Load the data}
setwd("C:/Users/marin/OneDrive/Documentos/GitHub/s223690/R")

de_genes <- read.delim("../data/de_genes.tsv", 
                         row.names = 1, 
                         stringsAsFactors = FALSE)
```

# Perform the Gene Ontology Enrichment Analysis

## Gene Ontology Enrichment Analysis focusing on the Celular Component

```{r GO enrichment CC}
#Perform the GO enrichment with a Biological Pathway pov
GO_results_CC <- enrichGO(gene = rownames(de_genes), OrgDb = "org.Hs.eg.db", keyType = "SYMBOL", ont = "CC")

#Filter depending on the gene count per pathway is not <5
GO_results_filtered_CC <- filter(GO_results_CC, Count >5)

#Transform the output into a data frame
GO_results_data_frame_CC <- as.data.frame(GO_results_filtered_CC)

#Plot
#jpeg("../results/GO_enrichment_CC.png", width = 600, height = 600, res = 100)

GO_results_CC <- plot(barplot(GO_results_filtered_CC,
                              showCategory = 10,
                              title = "GO of the DE genes Cellular Component",
                              font.size = 10))

#dev.off()
```

## Gene Ontology Enrichment Analysis focusing on the Biological Process

```{r GO enrichment BP}
#Perform the GO enrichment with a Biological Pathway pov
GO_results_BP <- enrichGO(gene = rownames(de_genes), OrgDb = "org.Hs.eg.db", keyType = "SYMBOL", ont = "BP")

#Filter depending on the gene count per pathway is not <5
GO_results_filtered_BP <- filter(GO_results_BP, Count >5)


#Transform the output into a data frame
GO_results_data_frame_BP <- as.data.frame(GO_results_filtered_BP)

#Plot
#jpeg("../results/GO_enrichment_BP.png", width = 600, height = 600, res = 100)

GO_results_BP <- plot(barplot(GO_results_filtered_BP, 
                                           showCategory = 10,
                                           title = "GO of the DE genes Biological Process",
                                           font.size = 10))

#dev.off()
```

## Gene Ontology Enrichment Analysis focusing on the Molecular Function

```{r GO enrichment MF}
#Perform the GO enrichment with a Biological Pathway pov
GO_results_MF <- enrichGO(gene = rownames(de_genes), OrgDb = "org.Hs.eg.db", keyType = "SYMBOL", ont = "MF")

#Filter depending on the gene count per pathway is not <5
GO_results_filtered_MF <- filter(GO_results_MF, Count >5)


#Transform the output into a data frame
GO_results_data_frame_MF <- as.data.frame(GO_results_filtered_MF)

#Plot
#jpeg("../results/GO_enrichment_MF.png", width = 600, height = 600, res = 100)

GO_results_MF <- plot(barplot(GO_results_filtered_MF, 
                              showCategory = 10,
                              title = "GO of the DE genes Molecular Function",                                        font.size = 10))
#dev.off()
```
